Major refactor, that now works compiling and linking wise.
Restructured all so that there are now
the base node: ZNode.
Then there is ConnectionNode, RecvNode and VariableGroupNode.
ConnectionNode matains a list of valid connections, that is connections in a certain state.
RecvNode, performs actual transmission and recieiving of raw data. It knows nothing about the state of open links such as connected.
VariableGroupNode matiains the variable groups that are linked to a connection.